+++ extra-src/Game/Halma/Configuration.hs
@@ -1,11 +1,14 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE KindSignatures #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE ExistentialQuantification #-}
 
 module Game.Halma.Configuration
-  ( NumberOfPlayers (..)
+  ( HalmaPlayers (..)
+  , twoPlayers, threePlayers, fourPlayers, fivePlayers, sixPlayers
   , getPlayers
   , Configuration (..)
   , defaultConfiguration
@@ -14,42 +17,54 @@ module Game.Halma.Configuration
 
 import Game.Halma.Board
 
-data NumberOfPlayers :: HalmaGridSize -> * where
-  TwoPlayers   :: NumberOfPlayers size
-  ThreePlayers :: NumberOfPlayers size
-  FourPlayers  :: NumberOfPlayers 'L
-  FivePlayers  :: NumberOfPlayers 'L
-  SixPlayers   :: NumberOfPlayers 'L
+data HalmaPlayers :: HalmaGridSize -> * -> * where
+  TwoPlayers   :: a -> a ->                     HalmaPlayers size a
+  ThreePlayers :: a -> a -> a ->                HalmaPlayers size a
+  FourPlayers  :: a -> a -> a -> a ->           HalmaPlayers 'L   a
+  FivePlayers  :: a -> a -> a -> a -> a ->      HalmaPlayers 'L   a
+  SixPlayers   :: a -> a -> a -> a -> a -> a -> HalmaPlayers 'L   a
 
-deriving instance Show (NumberOfPlayers size)
+deriving instance Eq a => Eq (HalmaPlayers size a)
+deriving instance Show a => Show (HalmaPlayers size a)
+deriving instance Functor (HalmaPlayers size)
+deriving instance Foldable (HalmaPlayers size)
+deriving instance Traversable (HalmaPlayers size)
 
-instance Eq (NumberOfPlayers size) where
-  a == b = show a == show b
+twoPlayers, threePlayers :: HalmaPlayers size ()
+twoPlayers = TwoPlayers () ()
+threePlayers = ThreePlayers () () ()
+fourPlayers, fivePlayers, sixPlayers :: HalmaPlayers 'L ()
+fourPlayers = FourPlayers () () () ()
+fivePlayers = FivePlayers () () () () ()
+sixPlayers = SixPlayers () () () () () ()
 
-getPlayers :: NumberOfPlayers size -> [Team]
-getPlayers numberOfPlayers =
-  case numberOfPlayers of
-    TwoPlayers   -> [North, South]
-    ThreePlayers -> [Northeast, South, Northwest]
-    FourPlayers  -> [Northeast, Southeast, Southwest, Northwest]
-    FivePlayers  -> [Northeast, Southeast, South, Southwest, Northwest]
-    SixPlayers   -> [minBound..maxBound]
+getPlayers :: HalmaPlayers size a -> [(Team, a)]
+getPlayers halmaPlayers =
+  case halmaPlayers of
+    TwoPlayers a b ->
+      [(North, a), (South, b)]
+    ThreePlayers a b c ->
+      [(Northeast, a), (South, b), (Northwest, c)]
+    FourPlayers a b c d ->
+      [(Northeast, a), (Southeast, b), (Southwest, c), (Northwest, d)]
+    FivePlayers a b c d e ->
+      [(Northeast, a), (Southeast, b), (South, c), (Southwest, d), (Northwest, e)]
+    SixPlayers a b c d e f ->
+      [(North, a), (Northeast, b), (Southeast, c), (South, d), (Southwest, e), (Northwest, f)]
 
-data Configuration :: HalmaGridSize -> * where
-  Configuration :: HalmaGrid size -> NumberOfPlayers size -> Configuration size
+data Configuration size a
+  = Configuration
+  { configurationGrid :: HalmaGrid size
+  , configurationPlayers :: HalmaPlayers size a
+  } deriving (Eq, Show, Functor, Foldable, Traversable)
 
-deriving instance Show (Configuration size)
+defaultConfiguration :: Configuration 'S ()
+defaultConfiguration = Configuration SmallGrid twoPlayers
 
-instance Eq (Configuration size) where
-  a == b = show a == show b
+data SomeConfiguration a = forall size. SomeConfiguration (Configuration size a)
 
-defaultConfiguration :: Configuration 'S
-defaultConfiguration = Configuration SmallGrid TwoPlayers
-
-data SomeConfiguration = forall size. SomeConfiguration (Configuration size)
-
-instance Show SomeConfiguration where
+instance Show a => Show (SomeConfiguration a) where
   show (SomeConfiguration a) = "SomeConfiguration (" ++ show a ++ ")"
 
-instance Eq SomeConfiguration where
+instance Show a => Eq (SomeConfiguration a) where
   SomeConfiguration a == SomeConfiguration b = show a == show b