+++ extra-src/Game/Halma/State.hs
@@ -10,20 +10,22 @@ import Game.TurnCounter
 
 import Data.Default (def)
 
-data HalmaState size
+data HalmaState size a
   = HalmaState
   { hsRuleOptions :: RuleOptions
   , hsBoard :: HalmaBoard size
-  , hsTurnCounter :: TurnCounter Team
+  , hsTurnCounter :: TurnCounter (Team, a)
   , hsLastMoved :: Maybe (Int, Int)
   } deriving (Eq, Show)
 
-newGame :: Configuration size -> HalmaState size
+newGame :: Configuration size a -> HalmaState size a
 newGame (Configuration halmaGrid nop) =
   HalmaState
-  { hsRuleOptions = def
-  , hsBoard = initialBoard halmaGrid (flip elem players)
-  , hsTurnCounter = newTurnCounter players
-  , hsLastMoved = Nothing
-  }
-  where players = getPlayers nop
+    { hsRuleOptions = def
+    , hsBoard = initialBoard halmaGrid isActive
+    , hsTurnCounter = newTurnCounter players
+    , hsLastMoved = Nothing
+    }
+  where
+    players = getPlayers nop
+    isActive color = color `elem` (fst <$> players)