+++ extra-src/Game/Halma/TelegramBot/Types.hs
@@ -1,82 +1,84 @@
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE StandaloneDeriving #-}
 
 module Game.Halma.TelegramBot.Types
   ( Player (..)
+  , showUser
   , showPlayer
-  , MatchResult (..)
+  , GameResult (..)
   , Match (..)
   , newMatch
   , MatchState (..)
   , ChatId
+  , PlayersSoFar (..)
   , BotState (..)
   , initialBotState
-  , CmdCall (..)
-  , parseCmdCall
   ) where
 
-import Game.Halma.Board
 import Game.Halma.Configuration
 import Game.Halma.State
 
-import Control.Monad (guard)
-import Data.Char (isAlphaNum, isSpace)
 import Data.Monoid ((<>))
 import qualified Data.Text as T
 import qualified Web.Telegram.API.Bot as TG
 
+-- TODO: remove this orphan instance
+deriving instance Eq TG.User
+
 data Player
   = AIPlayer
   | TelegramPlayer TG.User
-  deriving (Show)
+  deriving (Eq, Show)
+
+showUser :: TG.User -> T.Text
+showUser user =
+  case TG.user_username user of
+    Just username -> "@" <> username
+    Nothing ->
+      TG.user_first_name user <>
+      maybe "" (" " <>) (TG.user_last_name user)
 
 showPlayer :: Player -> T.Text
 showPlayer player =
   case player of
     AIPlayer -> "AI"
-    TelegramPlayer user ->
-      case TG.user_username user of
-        Just username -> username
-        Nothing ->
-          TG.user_first_name user <>
-          maybe "" (" " <>) (TG.user_last_name user)
+    TelegramPlayer user -> showUser user
 
-newtype MatchResult
-  = MatchResult
+newtype GameResult
+  = GameResult
   { numberOfMoves :: [(Player, Int)]
   } deriving (Show)
 
-data Match
+data Match size
   = Match
-  { matchPlayers :: [Player]
-  , matchHistory :: [MatchResult]
-  , matchCurrentGame :: Maybe (HalmaState 'S)
+  { matchConfig :: Configuration size Player
+  , matchHistory :: [GameResult]
+  , matchCurrentGame :: Maybe (HalmaState size Player)
   } deriving (Show)
 
-newMatch :: [Player] -> Either T.Text Match
-newMatch players =
-  let
-    config =
-      case length players of
-        i | i < 2 -> Left "can't start a match with at least two players!"
-        2 -> Right (Configuration SmallGrid TwoPlayers)
-        3 -> Right (Configuration SmallGrid ThreePlayers)
-        _ -> Left "can't start a match with more than three players!"
-  in
-    mkMatch <$> config
-  where
-    mkMatch config =
-      Match
-      { matchPlayers = players
-      , matchHistory = []
-      , matchCurrentGame = Just (newGame config)
-      }
+newMatch :: Configuration size Player -> Match size
+newMatch config =
+  Match
+    { matchConfig = config
+    , matchHistory = []
+    , matchCurrentGame = Just (newGame config)
+    }
+
+data PlayersSoFar a
+  = NoPlayers
+  | OnePlayer a
+  | forall size. EnoughPlayers (Configuration size a)
+
+deriving instance Show a => Show (PlayersSoFar a)
 
 data MatchState
   = NoMatch
-  | GatheringPlayers [Player]
-  | MatchRunning Match
-  deriving (Show)
+  | GatheringPlayers (PlayersSoFar Player)
+  | forall size. MatchRunning (Match size)
+
+deriving instance Show MatchState
 
 type ChatId = T.Text
 
@@ -95,22 +97,4 @@ initialBotState chatId token =
   , bsToken = token
   , bsNextId = 0
   , bsMatchState = NoMatch
-}
-
-data CmdCall
-  = CmdCall
-  { cmdCallName :: T.Text
-  , cmdCallArgs :: Maybe T.Text
-  } deriving (Show, Eq)
-
-parseCmdCall :: T.Text -> Maybe CmdCall
-parseCmdCall str = do
-  guard $ not (T.null str)
-  guard $ T.head str == '/'
-  let rest = T.tail str
-      isCmdChar c = isAlphaNum c || c `elem` ("_-" :: String)
-      (cmdName, rest') = T.span isCmdChar rest
-  guard $ not (T.null cmdName)
-  let rest'' = T.dropWhile isSpace rest'
-      args = if T.null rest'' then Nothing else Just rest''
-  pure $ CmdCall { cmdCallName = cmdName, cmdCallArgs = args }
+  }