+++ src/Game/Halma/AI/Competitive.hs
@@ -22,7 +22,7 @@ rate (t0, t1) board = rateTeam t0 board `against` rateTeam t1 board
     (Rating r0) `against` (Rating r1) = Rating (r0 - r1)
     _ `against` _ = error "unexpected team rating indicating loss"
 
-aiMove :: RuleOptions -> HalmaBoard size -> Perspective -> Move size
+aiMove :: RuleOptions -> HalmaBoard size -> Perspective -> Move
 aiMove opts board persp =
   snd $ prunedMinMaxSearch 3 opts board persp Nothing 
 
@@ -33,7 +33,7 @@ prunedMinMaxSearch
   -> HalmaBoard size
   -> Perspective
   -> Maybe Rating
-  -> (Rating, Move size)
+  -> (Rating, Move)
 prunedMinMaxSearch depth opts board persp mBound =
   go Nothing allOptions
   where
@@ -46,7 +46,7 @@ prunedMinMaxSearch depth opts board persp mBound =
         id
       else
         sortOn (flipRating . fst)
-    go :: Maybe (Rating, Move size) -> [(Rating, Move size)] -> (Rating, Move size)
+    go :: Maybe (Rating, Move) -> [(Rating, Move)] -> (Rating, Move)
     go Nothing (option@(rating, _move):options) =
       if isWin rating then
         option